+++ CHANGELOG.md
@@ -227,6 +227,9 @@
 - The XCache caching adapter has been deprecated as it is not compatible with the wildly
   deployed OPcache and does not perform better.
 
+ - The FirePhp logging adapter has been deprecated as it Firebug is not used by many
+   developers anymore.
+
 - The builtin mocking framework (`lithium\test\Mocker`) has been deprecated as alternatives
   exist and it is not needed as a core test dependency. This takes the task of maintaining full
   blown mocking from us. You might want to have a look at